requirements.txt
beautifulsoup4==4.3.2
Django==1.8
djangorestframework==3.1.1
ipython==3.1.0
requests==2.6.0
